% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_sd.R
\name{basic_sd}
\alias{basic_sd}
\title{Linear Steepest Descend}
\usage{
basic_sd(y, X, beta, tol = 0.001, maxit = 1000)
}
\arguments{
\item{y}{Response variable observations}

\item{X}{Covariates Matrix: each column contains observations for each covariate.}

\item{beta}{vector of initial parameters}

\item{tol}{Tolerance level for the optimization process, the default is 0.001.}

\item{maxit}{Maximum iterations number}
}
\value{
A list containing the fitted values for the beta vector and  the number of iterations performed
}
\description{
This function  computes the vector of parameters in a linear regression model via the Steepest Descend Method.
}
\examples{
set.seed(8675309)
n = 1000
x1 = rnorm(n)
x2 = rnorm(n)
y = 1 + .5*x1 + .2*x2 + rnorm(n)
X=cbind(x1,x2)
b_pre <- c(0,0,0)
basic_sd(y, X, b_pre)
}
