% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_pllLOOCV.R
\name{cm_pllLOOCV}
\alias{cm_pllLOOCV}
\title{Parallel Leave one out cross validation}
\usage{
cm_pllLOOCV(y, X, b_pre, tol = 0.001, maxit = 1000)
}
\arguments{
\item{y}{Response variable observations}

\item{X}{Covariates Matrix: each column contains observations for each covariate.}

\item{tol}{Tolerance level for the optimization process, the default is 0.001.}

\item{maxit}{Maximum iterations number}

\item{b:}{vector of initial parameters}
}
\value{
The mean squared error of prediction
}
\description{
This function performs a parallelized LOOCV to evaluate the prediction accuracy of a linear regression whose parameter are computed through a Steepest descent algorithm.
It always uses the maximum number of cores possible.
}
\examples{
set.seed(8675309)
n = 1000
x1 = rnorm(n)
x2 = rnorm(n)
y = 1 + .5*x1 + .2*x2 + rnorm(n)
X=cbind(x1,x2)
b_pre <- c(0,0,0)
cm_pllLOOCV(y, X, b_pre = c(0,0,0))
}
