% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_predict.R
\name{cm_predict}
\alias{cm_predict}
\title{Predict}
\usage{
cm_predict(X, beta)
}
\arguments{
\item{X}{matrix of covariates observations}

\item{beta}{vector of parameters of a linear regression model}
}
\value{
The mean squared error of prediction
}
\description{
This function computes the prediction
}
\examples{
set.seed(8675309)
n = 1000
x1 = rnorm(n)
x2 = rnorm(n)
y = 1 + .5*x1 + .2*x2 + rnorm(n)
X=cbind(x1,x2)
b_pre <- c(0,0,0)
beta <- basic_sd(y, X, b_pre)
cm_predict(X, beta)
cm_LOOCV(y, X, b_pre = c(0,0,0))
}
