% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate dataset for examples}
\usage{
generate_dataset(n)
}
\arguments{
\item{n:}{number of observations in the dataset}
}
\value{
A list containing the vector of observations for the dependent variable and a matrix containing the observations of the independent variables
}
\description{
This function generates the dataset used for examples
}
\examples{
set.seed(8675309)
generate_dataset(n)
}
