% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_gd_optim.R
\name{linear_gd_optim}
\alias{linear_gd_optim}
\title{Linear Gradient Descend}
\usage{
linear_gd_optim(b_pre, X, y, tolerance = 0.001, maxit = 1000, stepsize = 0.001)
}
\arguments{
\item{b_pre}{vector of initial parameters}

\item{X}{Covariates Matrix: each column contains observations for each covariate.}

\item{y}{Response variable observations}

\item{tolerance}{Tolerance level for the optimization process, the default is 0.001.}

\item{maxit}{Maximum iterations number}

\item{stepsize}{The value for the stepsize in the equation of the gradient descend}
}
\value{
A list containing the fitted values for the beta vector and the number of iterations performed
}
\description{
This function  computes the vector of parameters in a linear regression model via the Gradient Descend Method.
}
\examples{
set.seed(8675309)
n = 1000
x1 = rnorm(n)
x2 = rnorm(n)
y = 1 + .5*x1 + .2*x2 + rnorm(n)
X=cbind(x1,x2)
b_pre=c(0,0,0)
linear_gd_optim(b_pre,X,y)
}
