% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_gd_optim2.R
\name{linear_gd_optim2}
\alias{linear_gd_optim2}
\title{Illustration of crayon colors}
\usage{
linear_gd_optim2(
  b_pre,
  X,
  y,
  tolerance = 0.001,
  maxit = 1000,
  stepsize = 0.001
)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{cex}{character expansion for the text}

\item{mar}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{linear_gd_optim2}}
}
\examples{
plot_crayons()

}
