% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_gd_optim.R
\name{linear_gd_optim}
\alias{linear_gd_optim}
\title{Cacca}
\usage{
linear_gd_optim(b_pre, X, y, tolerance = 0.001, maxit = 1000, stepsize = 0.001)
}
\arguments{
\item{method2order}{method to order colors}

\item{cex}{character expansion for the text}

\item{mar}{margin parameters; vector of length 4}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{linear_gd_optim}}
}
\examples{
plot_crayons()

}
