% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_gd_optim2.R
\name{linear_gd_optim2}
\alias{linear_gd_optim2}
\title{Linear Gradient Descend (2)}
\usage{
linear_gd_optim2(
  b_pre,
  X,
  y,
  tolerance = 0.001,
  maxit = 1000,
  stepsize = 0.001
)
}
\arguments{
\item{b_pre}{vector of initial parameters}

\item{X}{Covariates Matrix: each column contains observations for each covariate.}

\item{y}{Response variable observations}

\item{tolerance}{Tolerance level for the optimization process, the default is 0.001.}

\item{maxit}{Maximum iterations number. Default is 1000.}

\item{stepsize}{The value for the stepsize in the equation of the gradient descend. Default is 0.001.}
}
\value{
A list containing the fitted values for the beta vector and the number of iterations performed.
}
\description{
This function is the slower version of \code{\link{linear_gd_optim}} given that it makes use of the \code{\link{grad}} function contained in the package "numDeriv".
}
\examples{
set.seed(8675309)
n = 1000
x1 = rnorm(n)
x2 = rnorm(n)
y = 1 + .5*x1 + .2*x2 + rnorm(n)
X=cbind(x1,x2)
b_pre=c(0,0,0)
linear_gd_optim(b_pre,X,y)#'
}
